+++ src/test/java/be/ugent/rml/Custom_RML_FnO_Mapper_Test.java
@@ -2,58 +2,57 @@ package be.ugent.rml;
 
 import be.ugent.rml.functions.FunctionLoader;
 import be.ugent.rml.functions.FunctionUtils;
+import be.ugent.rml.store.QuadStore;
+import be.ugent.rml.store.RDF4JStore;
+import org.eclipse.rdf4j.rio.RDFFormat;
 import org.junit.Test;
 
 import java.io.File;
-import java.io.IOException;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 
 public class Custom_RML_FnO_Mapper_Test extends TestFunctionCore {
 
     @Test
-    public void evaluate_A001() {
-        try {
-            File myFile = new File("./src/test/resources/rml-fno-test-cases/functions_dynamic.ttl");
-            FunctionLoader functionLoader = new FunctionLoader(myFile);
-            Executor executor = this.createExecutor("./rml-fno-test-cases/RMLFNOTC0001-CSV/mapping.ttl", functionLoader);
-            doMapping(executor, "./rml-fno-test-cases/RMLFNOTC0001-CSV/output.ttl");
-            assertTrue(functionLoader.getLibraryPath("GrelFunctions").endsWith("GrelFunctions_dynamic.jar"));
-        } catch (Exception e) {
-            logger.error(e.getMessage(), e);
-            fail();
-        }
+    public void evaluate_A001() throws Exception {
+        QuadStore functionDescriptionTriples = new RDF4JStore();
+        functionDescriptionTriples.read(Utils.getInputStreamFromFile(new File("./src/test/resources/rml-fno-test-cases/functions_dynamic.ttl")), null, RDFFormat.TURTLE);
+        FunctionLoader functionLoader = new FunctionLoader(functionDescriptionTriples);
+        Executor executor = this.createExecutor("./rml-fno-test-cases/RMLFNOTC0001-CSV/mapping.ttl", functionLoader);
+        doMapping(executor, "./rml-fno-test-cases/RMLFNOTC0001-CSV/output.ttl");
+        assertTrue(functionLoader.getLibraryPath("GrelFunctions").endsWith("GrelFunctions_dynamic.jar"));
     }
 
     @Test
-    public void evaluate_A002() {
+    public void evaluate_A002() throws Exception {
         Executor executor = doPreloadMapping("./rml-fno-test-cases/RMLFNOTC0001-CSV/mapping.ttl", "./rml-fno-test-cases/RMLFNOTC0001-CSV/output.ttl");
-        assertEquals(executor.getFunctionLoader().getLibraryPath("GrelFunctions"), "__local");
+        assertEquals("__local", executor.getFunctionLoader().getLibraryPath("io.fno.grel.ArrayFunctions"));
     }
 
     @Test
-    public void evaluate_A003() {
-        Executor executor = doMapping("./rml-fno-test-cases/RMLFNOTC0001-CSV/mapping.ttl", "./rml-fno-test-cases/RMLFNOTC0001-CSV/output.ttl");
-        String libPath = executor.getFunctionLoader().getLibraryPath("GrelFunctions");
-        Class cls = null;
-        try {
-            cls = FunctionUtils.functionRequire(new File(libPath), "GrelFunctions");
-        } catch (IOException e) {
-            logger.error(e.getMessage(), e);
-            fail();
-        }
+    public void evaluate_A003() throws Exception {
+        QuadStore functionDescriptionTriples = new RDF4JStore();
+        functionDescriptionTriples.read(Utils.getInputStreamFromFile(new File("./src/test/resources/rml-fno-test-cases/functions_dynamic.ttl")), null, RDFFormat.TURTLE);
+        FunctionLoader functionLoader = new FunctionLoader(functionDescriptionTriples);
+        // You first need to execute the mapping, bc the libraryMap of loaded Jars is dynamically built
+        Executor executor = this.createExecutor("./rml-fno-test-cases/RMLFNOTC0001-CSV/mapping.ttl", functionLoader);
+        doMapping(executor, "./rml-fno-test-cases/RMLFNOTC0001-CSV/output.ttl");
+        String libPath = functionLoader.getLibraryPath("GrelFunctions");
+        Class cls = FunctionUtils.functionRequire(new File(libPath), "GrelFunctions");
         assertEquals("GrelFunctions", cls.getName());
     }
 
+    /**
+     * Tests whether the function `idlab-fn:trueCondition` is supported correctly by the mapper
+     */
     @Test
-    public void evaluate_A004() {
+    public void evaluate_A004() throws Exception {
         doPreloadMapping("./rml-fno-test-cases/RMLFNOTCA004/mapping.ttl", "./rml-fno-test-cases/RMLFNOTCA004/output.ttl");
     }
 
     @Test
-    public void evaluate_A004b() {
+    public void evaluate_A004b() throws Exception {
         doPreloadMapping("./rml-fno-test-cases/RMLFNOTCA004b/mapping.ttl", "./rml-fno-test-cases/RMLFNOTCA004b/output.ttl");
     }
 
@@ -61,7 +60,15 @@ public class Custom_RML_FnO_Mapper_Test extends TestFunctionCore {
      * Tests whether the function idlab-fn:readFile is supported correctly by the mapper
      */
     @Test
-    public void evaluate_A005() {
+    public void evaluate_A005() throws Exception {
         doPreloadMapping("./rml-fno-test-cases/RMLFNOTCA005/mapping.ttl", "./rml-fno-test-cases/RMLFNOTCA005/output.ttl");
+    }
+
+    /**
+     * Tests whether the function idlab-fn:inRange is supported correctly by the mapper
+     */
+    @Test
+    public void evaluate_A006() {
+        doMapping("./rml-fno-test-cases/RMLFNOTCA006/mapping.ttl", "./rml-fno-test-cases/RMLFNOTCA006/output.ttl");
     }
 }