+++ src/test/java/be/ugent/rml/readme/ReadmeTest.java
@@ -3,7 +3,6 @@ package be.ugent.rml.readme;
 import be.ugent.rml.Executor;
 import be.ugent.rml.Utils;
 import be.ugent.rml.functions.FunctionLoader;
-import be.ugent.rml.functions.lib.GrelProcessor;
 import be.ugent.rml.functions.lib.IDLabFunctions;
 import be.ugent.rml.records.RecordsFactory;
 import be.ugent.rml.store.QuadStore;
@@ -36,10 +35,9 @@ public class ReadmeTest {
 
             // Set up the functions used during the mapping
             Map<String, Class> libraryMap = new HashMap<>();
-            libraryMap.put("GrelFunctions", GrelProcessor.class);
             libraryMap.put("IDLabFunctions", IDLabFunctions.class);
 
-            FunctionLoader functionLoader = new FunctionLoader(null, null, libraryMap);
+            FunctionLoader functionLoader = new FunctionLoader(null, libraryMap);
 
             // Set up the outputstore (needed when you want to output something else than nquads
             QuadStore outputStore = new RDF4JStore();